feat: merge to master

+++ apps/web/src/common/modules/favorites/favorite-button/FavoriteButton.vue
@@ -1,5 +1,6 @@
 <template>
     <p-i
+        v-if="!isAdminMode"
         v-show="(visibleActiveCaseOnly || readOnly) ? active : true"
         :name="active ? 'ic_favorite-filled': 'ic_favorite'"
         width="1rem"
@@ -22,6 +23,8 @@ import { PI } from '@spaceone/design-system';
 
 import { store } from '@/store';
 
+import { useAppContextStore } from '@/store/app-context/app-context-store';
+import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-workspace-store';
 import type { FavoriteConfig } from '@/store/modules/favorite/type';
 import { FAVORITE_TYPE_TO_STATE_NAME } from '@/store/modules/favorite/type';
 
@@ -53,13 +56,18 @@ export default defineComponent<FavoriteButtonProps>({
         },
     },
     setup(props) {
+        const userWorkspaceStore = useUserWorkspaceStore();
+        const appContextStore = useAppContextStore();
         const state = reactive({
+            isAdminMode: computed(() => appContextStore.getters.isAdminMode),
+            currentWorkspaceId: computed(() => userWorkspaceStore.getters.currentWorkspaceId),
             isLoading: computed(() => store.state.favorite.isLoading[props.favoriteType]),
             hasLoaded: computed(() => Array.isArray(state.favoriteItems)),
             favoriteItems: computed<FavoriteConfig[]|null>(() => {
                 const stateName = FAVORITE_TYPE_TO_STATE_NAME[props.favoriteType];
                 if (!stateName) return [];
-                return store.state.favorite[stateName];
+                const items = store.state.favorite[stateName];
+                return items ? store.state.favorite[stateName].filter((item) => item.workspaceId === state.currentWorkspaceId) : null;
             }),
             favoriteItemMap: computed<Record<string, FavoriteConfig>>(() => {
                 const result: Record<string, FavoriteConfig> = {};
@@ -74,15 +82,18 @@ export default defineComponent<FavoriteButtonProps>({
         });
 
         const handleClickFavoriteButton = async (event: MouseEvent) => {
+            if (state.isAdminMode) return;
             event.stopPropagation();
             if (props.readOnly) return;
             if (state.favoriteItemMap[props.itemId]) {
                 await store.dispatch('favorite/removeItem', {
+                    workspaceId: state.currentWorkspaceId,
                     itemId: props.itemId,
                     itemType: props.favoriteType,
                 });
             } else {
                 await store.dispatch('favorite/addItem', {
+                    workspaceId: state.currentWorkspaceId,
                     itemId: props.itemId,
                     itemType: props.favoriteType,
                 });
